+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import Router from "./router";
+import Router from "./use/router";
 import { store } from "./use/store/store";
 import "./assets/styles/index.css";
