+++ src/use/router/index.ts
@@ -1,16 +1,16 @@
 import { createRouter, createWebHistory } from "vue-router";
-import LoginView from "../views/common/Login.vue";
-import StudentCourseView from "../views/student/StudentCourseList.vue";
-import LecturerCourseView from "../views/lecturer/LecturerCourseList.vue";
-import AdminAccountListView from "../views/admin/AdminAccountList.vue";
-import CourseFormSuspenseWrapper from "../views/lecturer/EditCreateCourseForm.vue";
-import AccountFormSuspenseWrapper from "../views/admin/EditCreateAccountForm.vue";
-import Redirect from "../views/errors/403.vue";
-import ProfileWrapper from "../components/profile/Wrapper.vue";
-import Settings from "../views/common/Settings.vue";
-import PageNotFound from "../views/errors/404.vue";
-import WelcomePage from "../views/common/Welcome.vue";
-import AboutPage from "../views/common/About.vue";
+import LoginView from "@/views/common/Login.vue";
+import StudentCourseView from "@/views/student/StudentCourseList.vue";
+import LecturerCourseView from "@/views/lecturer/LecturerCourseList.vue";
+import AdminAccountListView from "@/views/admin/AdminAccountList.vue";
+import CourseFormSuspenseWrapper from "@/views/lecturer/EditCreateCourseForm.vue";
+import AccountFormSuspenseWrapper from "@/views/admin/EditCreateAccountForm.vue";
+import Redirect from "@/views/errors/403.vue";
+import ProfileWrapper from "@/components/profile/Wrapper.vue";
+import Settings from "@/views/common/Settings.vue";
+import PageNotFound from "@/views/errors/404.vue";
+import WelcomePage from "@/views/common/Welcome.vue";
+import AboutPage from "@/views/common/About.vue";
 import { checkPrivilege } from "@/use/PermissionHelper";
 import { Role } from "@/entities/Role";
 import { useStore } from "@/use/store/store";
