+++ src/views/lecturer/PrivateProfile.vue
@@ -28,9 +28,7 @@
                 <div class="lg:flex">
                     <div class="w-full lg:w-1/3 lg:block mr-12 flex flex-col mb-4">
                         <label class="block text-gray-700 text-lg font-medium mb-2">Notifications</label>
-                        <label class="block text-gray-600">
-                            Control when and how to stay updated about recent activities.
-                        </label>
+                        <label class="block text-gray-600"> Control when and how to stay updated about recent activities. </label>
                     </div>
                     <div class="w-full lg:w-2/3">
                         <div class="mb-6 flex items-start">
@@ -60,7 +58,7 @@
     import AddressSection from "@/components/profile/AddressSection.vue";
     import ResearchSection from "@/components/profile/lecturer/ResearchSection.vue";
     import { ref } from "vue";
-    import Router from "@/router";
+    import Router from "@/use/router";
     import UserManagement from "@/api/UserManagement";
     import Lecturer from "../../api/api_models/user_management/Lecturer";
     import { useModelWrapper } from "@/use/ModelWrapper";
